% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDhsData.R
\name{getDHSdata}
\alias{getDHSdata}
\title{Get your survival data ready}
\usage{
getDHSdata(iso, survyear, vaccine, vax_age = 3, assessment = FALSE)
}
\arguments{
\item{iso}{country specific iso code (eg. "GH")}

\item{survyear}{vector, year(s) of survey}

\item{vaccine}{character, options: MCV1, MCV2}

\item{vax_age}{integer, mininum vaccination age allowed}

\item{assessment}{boolean, TRUE to print assessment report of the data}
}
\value{
A nicely formatted dataframe for survival analysis
}
\description{
Get your survival data ready
}
