% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFit.R
\name{survivalFit}
\alias{survivalFit}
\title{Fit a boubly-censored survival model}
\usage{
survivalFit(datainput, plotoutput = FALSE, xlab = "", ylab = "", ...)
}
\arguments{
\item{datainput}{a data.frame similar to \code{getDhsdata()}}

\item{plotoutput}{logical, default is FALSE}

\item{xlab}{character, label for x-axis}

\item{ylab}{character, label for y-axis}

\item{...}{additional arguments for \code{plot()}}
}
\value{
list, cumulative probabilities with confidence intervals
}
\description{
Fit a boubly-censored survival model
}
