% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeliness-package.R
\docType{package}
\name{vaccinetimeliness-package}
\alias{vaccinetimeliness}
\alias{vaccinetimeliness-package}
\title{vaccinetimeliness: Put your DHS data in a `Survival ready` format}
\description{
`vaccinetimeliness` helps get your DHS data in a format that is appropriate for survival analysis. Additionally, helper functions allow for easy modelling of survival data in the presense of doubly censored data (both left and right censoring).
}
\author{
\strong{Maintainer}: Abraham Adokwei \email{abrahamaddo14@gmail.com} (\href{https://orcid.org/0009-0000-7044-2902}{ORCID})

Authors:
\itemize{
  \item Abraham Adokwei \email{abrahamaddo14@gmail.com} (\href{https://orcid.org/0009-0000-7044-2902}{ORCID})
  \item Amy Winter \email{awinter@uga.edu}
}

}
\keyword{internal}
